+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Add `ConfigServerClient` feign client to fetch configuration remotely. (#446)
 - Add `ConfigServerClientBuilder` helper class to create feign client instances. (#447)
+- Rename `blockchainAdapterUrl` to `configServerUrl` in class `PublicConfiguration`. (#451)
 
 ### Quality